requirements.txt
pyTelegramBotAPI==4.17.0
gspread>=6.0.0
google-auth>=2.0.0

